requirements.txt
alembic==0.9.9
awscli==1.16.176
bcrypt==3.1.6
blinker==1.4
boto3==1.9.166
botocore==1.12.166
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
click==6.7
colorama==0.3.9
docutils==0.14
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-Scss==0.5
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
google-api-core==1.11.1
google-auth==1.6.3
google-cloud-core==1.0.1
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
gunicorn==19.9.0
httplib2==0.11.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.4
lazy==1.3
Mako==1.0.7
MarkupSafe==1.0
oauth2==1.9.0.post1
oauthlib==2.0.7
Pillow==6.0.0
protobuf==3.8.0
psycopg2-binary==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pyScss==1.3.5
python-dateutil==2.7.2
python-dotenv==0.10.2
python-editor==1.0.3
pytz==2019.1
PyYAML==3.13
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.2.1
six==1.11.0
SQLAlchemy==1.2.6
SQLAlchemy-Utils==0.33.2
urllib3==1.22
URLObject==2.4.3
Werkzeug==0.14.1
wincertstore==0.2
WTForms==2.1
